LimitOrderBook.jl

LimitOrderBook.AcctMap — Type

AcctMap{Oid,Aid,ST,PT}

Collection of open orders by account.

{Oid,Aid,ST,PT} characterize the type of Order present in the AcctMap. See documentation on Order for more information on these types.

The account map has the structure of a nested Dict. The outer key is the account id, mapping to an AVLTree of Orders keyed by order id.

LimitOrderBook.Order — Type

Order{Oid<:Integer,Aid<:Integer,ST<:Real,PT<:Real}

Type representing a limit order.

An Order{Oid<:Integer,Aid<:Integer,ST<:Real,PT<:Real} is a struct representing a Limit Order which contains

An unique order id orderid::Oid,

A side, side::Symbol, where it is expected that side=:ASK or side=:BID to represent the side of the book the order will rest in.

An order size, size::ST

A price the order is set at, price::PT and,

Optionally, a unique account ID acctid::Union{Aid,Nothing}, which is set to nothing if the account is unknown or irrelevant.

One can create a new Order as Order{Oid,Aid,ST,PT}(orderid,side,size,price[,acctid=nothing]), where the types of size and price will be cast to the correct types. The orderid and acctid types will not be cast to avoid ambiguity.

LimitOrderBook.OrderBook — Type

OrderBook{Oid,Aid,ST,PT}

Object representing a Limit Order Book.

An OrderBook{Oid,Aid,ST,PT} is a data structure containing limit orders represented as objects of type Order{Oid,Aid,ST,PT}. See documentation on Order for more information on these parametric types.

Initialize an empty limit order book with OrderBook{Oid,Aid,ST,PT}()

LimitOrderBook.best_bid_ask — Method

best_bid_ask(ob::OrderBook)

Return best bid/ask prices in order book as a Tuple

LimitOrderBook.book_depth_info — Method

book_depth_info(ob::OrderBook; max_depth=5)

Returns prices, volumes and order counts at bid and ask in ob::OrderBook until fixed depth max_depth as a nested Dict.

LimitOrderBook.cancel_limit_order! — Method

cancel_limit_order!(ob::OrderBook, o::Order)

Cancels order o::Order from LimitOrderBook.

LimitOrderBook.cancel_limit_order! — Method

cancel_limit_order!(ob::OrderBook, orderid, price, side[, acct_id=nothing])

Cancels order with matching information from OrderBook.

Provide acct_id if known to guarantee correct account tracking.

LimitOrderBook.get_acct — Method

get_acct(ob::OrderBook{Oid,Aid,ST,PT},acct_id::Aid)

Return all open orders assigned to account acct_id

LimitOrderBook.n_orders_bid_ask — Method

n_orders_bid_ask(ob::OrderBook)

Return total number of orders on each side of order book, returned as a Tuple

LimitOrderBook.order_types — Method

order_types(::X{Oid,Aid,ST,PT})

Return parametric types of either an Order, OrderQueue, OneSidedbook or OrderBook.

Example

order_types(Order{Int64,Int64,Int64,Float64}) = (Int64,Int64,Int64,Float64)

LimitOrderBook.submit_limit_order! — Method

submit_limit_order!(ob::OrderBook, orderid, price, size, side[, acct_id=nothing])

Enter limit order with matching properties into ob::OrderBook.

If acct_id provided, account holdings are tracked in ob.AcctMap.

LimitOrderBook.submit_market_order! — Method

submit_market_order!(ob::OrderBook{Oid,Aid,ST,PT},side::Symbol,mo_size)

Submit market order to OrderBook on given side (:ASK or :BID) and size mo_size.

Market orders are filled by price-time priority.

Returns tuple ( ord_lst::Vector{Order}, complete_status::Symbol, left_to_trade::ST ) where

ord_lst is a list of LOs that MO matched with

complete_status is either :COMPLETE or :INCOMPLETE, denoting whether there was enough liquidity to complete the order or not.

left_to_trade is the remaining size of un-filled order ( ==0 if :COMPLETE, >0 if :INCOMPLETE)

LimitOrderBook.volume_bid_ask — Method

volume_bid_ask(ob::OrderBook)

Return total bid and ask volume from order book as a Tuple.

LimitOrderBook.write_csv — Method

write_csv(
    io::IO,
    ob::OrderBook;
    row_formatter = _order_to_csv,
    header = "TRD,ID,SIDE,SIZE,PX,ACCT",
    )

Write OrderBook ob to an IO stream into csv format where each row corresponds to an order The formatting for each row is given by the function argument row_formatter(::Order)::String. The csv header can be provided as an argument where setting it to nothing writes no header.