LimitOrderBook.jl

LimitOrderBook.AcctMap — Type

AcctMap{Oid,Aid,ST,PT}

Collection of open orders by account.

{Oid,Aid,ST,PT} characterize the type of Order present in the AcctMap. See documentation on Order for more information on these types.

The account map has the structure of a nested Dict. The outer key is the account id, mapping to an AVLTree of Orders keyed by order id.

LimitOrderBook.Order — Type

Order{Oid<:Integer,Aid<:Integer,ST<:Real,PT<:Real}

Type representing a limit order.

An Order{Oid<:Integer,Aid<:Integer,ST<:Real,PT<:Real} is a struct representing a Limit Order which contains

An unique order id orderid::Oid,

side::Symbol, the side of the book the order will rest in, where either side=:ASK or side=:BID.

size::ST, the order size

price::PT, the price the order is set at

Optionally, a unique account ID acctid::Union{Aid,Nothing}, which is set to nothing if the account is unknown or irrelevant.

One can create a new Order as Order{Oid,Aid,ST,PT}(orderid, side, size, price [,acctid=nothing]), where the types of size and price will be cast to the correct types. The orderid and acctid types will not be cast in order to avoid ambiguity.

LimitOrderBook.OrderBook — Type

OrderBook{Oid,Aid,ST,PT}

An OrderBook is a data structure containing limit orders represented as objects of type Order{Oid,Aid,ST,PT}.

See documentation on Order for more information on the parametric type Order{Oid,Aid,ST,PT}.

How to use Orderbook:

Initialize an empty limit order book as OrderBook{Oid,Aid,ST,PT}()

Submit or cancel limit orders with submit_limit_order! and cancel_limit_order!.

Submit market orders with submit_market_order!

Retrieve order book state information with print or show methods, as well as book_depth_info, best_bid_ask, volume_bid_ask, n_orders_bid_ask and get_acct

Write book state to csv file with write_csv.

LimitOrderBook.best_bid_ask — Method

best_bid_ask(ob::OrderBook)

Return best bid/ask prices in order book as a Tuple

LimitOrderBook.book_depth_info — Method

book_depth_info(ob::OrderBook; max_depth=5)

Returns prices, volumes and order counts at bid and ask in ob::OrderBook until fixed depth max_depth as a nested Dict.

LimitOrderBook.cancel_limit_order! — Method

cancel_limit_order!(ob::OrderBook, o::Order)

Cancels order o::Order from LimitOrderBook.

LimitOrderBook.cancel_limit_order! — Method

cancel_limit_order!(ob::OrderBook, orderid, price, side [, acct_id=nothing])

Cancels order with matching information from OrderBook.

Provide acct_id if known to guarantee correct account tracking.

LimitOrderBook.get_acct — Method

get_acct(ob::OrderBook{Oid,Aid,ST,PT},acct_id::Aid)

Return all open orders assigned to account acct_id

LimitOrderBook.n_orders_bid_ask — Method

n_orders_bid_ask(ob::OrderBook)

Return total number of orders on each side of order book, returned as a Tuple

LimitOrderBook.order_types — Method

order_types(::X{Oid,Aid,ST,PT})

Return parametric types of either an Order, OrderQueue, OneSidedbook or OrderBook.

Example

order_types(Order{Int64,Int64,Int64,Float64}) = (Int64,Int64,Int64,Float64)

LimitOrderBook.submit_limit_order! — Method

submit_limit_order!(ob::OrderBook, orderid, price, size, side [, acct_id=nothing])

Enter limit order with matching properties into ob::OrderBook.

If acct_id provided, account holdings are tracked in ob.AcctMap.

LimitOrderBook.submit_market_order! — Method

submit_market_order!(ob::OrderBook{Oid,Aid,ST,PT},side::Symbol,mo_size)

Submit market order to OrderBook on given side (:ASK or :BID) and size mo_size.

Market orders are filled by price-time priority.

Returns tuple ( ord_lst::Vector{Order}, complete_status::Symbol, left_to_trade::ST ) where

ord_lst is a list of limit orders that market order matched with

left_to_trade is the remaining size of un-filled order ( ==0 if order is complete, >0 if incomplete)

LimitOrderBook.volume_bid_ask — Method

volume_bid_ask(ob::OrderBook)

Return total bid and ask volume from order book as a Tuple.

LimitOrderBook.write_csv — Method

write_csv(
    io::IO,
    ob::OrderBook;
    row_formatter = _order_to_csv,
    header = "TRD,ID,SIDE,SIZE,PX,ACCT",
    )

Write OrderBook ob to an IO stream into csv format where each row corresponds to an order The formatting for each row is given by the function argument row_formatter(::Order)::String. The csv header can be provided as an argument where setting it to nothing writes no header.